+++ CHANGELOG.md
@@ -1,8 +1,7 @@
 # avro_schema_registry-client
 
 ## v0.2.0
-- Add `register_and_lookup` and `register_without_lookup` methods to
-  `AvroSchemaRegistry::Client`.
+- Add `register_without_lookup` methods to `AvroSchemaRegistry::Client`.
 
 ## v0.1.0
 - Initial version